// app-client/src/graphql.ts
export type Maybe<T> = T | null | undefined;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  date: any;
  inet: any;
  json: any;
  smallint: any;
  time: any;
  timestamptz: any;
  uuid: any;
};
// ...
export type LoginOutput = {
  __typename?: 'LoginOutput';
  /** An object relationship */
  customer: Customers;
  customer_id: Maybe<Scalars['uuid']>;
  session_token: Scalars['String'];
};
// ...
export type Customers = {
  __typename?: 'customers';
  contact_name: Maybe<Scalars['String']>;
  created_at: Scalars['timestamptz'];
  firebase_uid: Scalars['String'];
  id: Scalars['uuid'];
  /** An object relationship */
  membership: Maybe<Memberships>;
  /** An array relationship */
  personas: Array<Personas>;
  /** An aggregate relationship */
  personas_aggregate: Personas_Aggregate;
  stripe_customer_id: Maybe<Scalars['String']>;
  unverified_email: Scalars['String'];
  verified_email: Maybe<Scalars['String']>;
};
// ...
export type Membership_Levels = {
  __typename?: 'membership_levels';
  value: Scalars['String'];
};
// ...
export enum Membership_Levels_Enum {
  Basic = 'basic',
  Patron = 'patron',
  Supporter = 'supporter'
}
// ...
export type Memberships = {
  __typename?: 'memberships';
  active_until: Maybe<Scalars['timestamptz']>;
  created_at: Scalars['timestamptz'];
  /** An object relationship */
  customer: Customers;
  id: Scalars['uuid'];
  level: Membership_Levels_Enum;
  lifetime: Scalars['Boolean'];
  /** An object relationship */
  membership_level: Membership_Levels;
  sabbatical_available_until: Maybe<Scalars['timestamptz']>;
  stripe_subscription_id: Maybe<Scalars['String']>;
  stripe_subscription_status: Maybe<Scalars['String']>;
  updated_at: Maybe<Scalars['timestamptz']>;
};
// ...
export type Personas = {
  __typename?: 'personas';
  created_at: Scalars['timestamptz'];
  /** An object relationship */
  customer: Customers;
  customer_id: Scalars['uuid'];
  id: Scalars['uuid'];
  list_address_with_byline: Scalars['Boolean'];
  name: Scalars['String'];
  photo_url: Maybe<Scalars['String']>;
  preferred_broadcast_time: Scalars['time'];
  updated_at: Maybe<Scalars['timestamptz']>;
};

/** aggregated selection of "personas" */
export type Personas_Aggregate = {
  __typename?: 'personas_aggregate';
  aggregate: Maybe<Personas_Aggregate_Fields>;
  nodes: Array<Personas>;
};
// ...
export type LoginMutationVariables = Exact<{ [key: string]: never; }>;


export type LoginMutation = (
  { __typename?: 'mutation_root' }
  & { login: Maybe<(
    { __typename?: 'LoginOutput' }
    & Pick<LoginOutput, 'session_token'>
    & { customer: (
      { __typename?: 'customers' }
      & CustomerDetailsFragment
    ) }
  )> }
);

export type CustomerDetailsQueryVariables = Exact<{
  customerId: Scalars['uuid'];
}>;


export type CustomerDetailsQuery = (
  { __typename?: 'query_root' }
  & { customer: Maybe<(
    { __typename?: 'customers' }
    & CustomerDetailsFragment
  )> }
);

export type CustomerDetailsFragment = (
  { __typename?: 'customers' }
  & Pick<Customers, 'id' | 'contact_name' | 'unverified_email' | 'verified_email'>
  & { membership: Maybe<(
    { __typename?: 'memberships' }
    & Pick<Memberships, 'id' | 'active_until' | 'level' | 'lifetime' | 'sabbatical_available_until'>
  )>, personas: Array<(
    { __typename?: 'personas' }
    & Pick<Personas, 'id' | 'list_address_with_byline' | 'name' | 'photo_url' | 'preferred_broadcast_time'>
  )> }
);
// ...
export const CustomerDetailsFragmentDoc: DocumentNode<CustomerDetailsFragment, unknown> = {"kind":"Document","definitions":[{"kind":"FragmentDefinition","name":{"kind":"Name","value":"customerDetails"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"customers"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"contact_name"}},{"kind":"Field","name":{"kind":"Name","value":"membership"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"active_until"}},{"kind":"Field","name":{"kind":"Name","value":"level"}},{"kind":"Field","name":{"kind":"Name","value":"lifetime"}},{"kind":"Field","name":{"kind":"Name","value":"sabbatical_available_until"}}]}},{"kind":"Field","name":{"kind":"Name","value":"personas"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"list_address_with_byline"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"photo_url"}},{"kind":"Field","name":{"kind":"Name","value":"preferred_broadcast_time"}}]}},{"kind":"Field","name":{"kind":"Name","value":"unverified_email"}},{"kind":"Field","name":{"kind":"Name","value":"verified_email"}}]}}]};
export const LoginDocument: DocumentNode<LoginMutation, LoginMutationVariables> = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"mutation","name":{"kind":"Name","value":"Login"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"login"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"session_token"}},{"kind":"Field","name":{"kind":"Name","value":"customer"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"customerDetails"}}]}}]}}]}},...CustomerDetailsFragmentDoc.definitions]};
export const CustomerDetailsDocument: DocumentNode<CustomerDetailsQuery, CustomerDetailsQueryVariables> = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"CustomerDetails"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"customerId"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"uuid"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"customer"},"name":{"kind":"Name","value":"customers_by_pk"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"customerId"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"customerDetails"}}]}}]}},...CustomerDetailsFragmentDoc.definitions]};